requirements.txt
requests
beautifulsoup4
transformers
torch
faiss-cpu
sentence-transformers
python-dotenv
